Issue #12341 - only send QpackEncoder instructions if change in capacity
