+++ jetty-core/jetty-http3/jetty-http3-qpack/src/test/java/org/eclipse/jetty/http3/qpack/SectionAcknowledgmentTest.java
@@ -17,14 +17,16 @@ import java.nio.ByteBuffer;
 
 import org.eclipse.jetty.http3.qpack.QpackException.SessionException;
 import org.eclipse.jetty.http3.qpack.internal.instruction.SectionAcknowledgmentInstruction;
+import org.eclipse.jetty.http3.qpack.util.TestDecoderHandler;
+import org.eclipse.jetty.http3.qpack.util.TestEncoderHandler;
 import org.eclipse.jetty.util.BufferUtil;
 import org.eclipse.jetty.util.NanoTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.eclipse.jetty.http3.qpack.QpackTestUtil.encode;
-import static org.eclipse.jetty.http3.qpack.QpackTestUtil.toBuffer;
-import static org.eclipse.jetty.http3.qpack.QpackTestUtil.toMetaData;
+import static org.eclipse.jetty.http3.qpack.util.QpackTestUtil.encode;
+import static org.eclipse.jetty.http3.qpack.util.QpackTestUtil.toBuffer;
+import static org.eclipse.jetty.http3.qpack.util.QpackTestUtil.toMetaData;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
