+++ jetty-core/jetty-http3/jetty-http3-qpack/src/test/java/org/eclipse/jetty/http3/qpack/util/TestEncoderHandler.java
@@ -11,24 +11,17 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.http3.qpack;
+package org.eclipse.jetty.http3.qpack.util;
 
 import java.nio.ByteBuffer;
 import java.util.LinkedList;
 import java.util.List;
 
-import org.eclipse.jetty.http.MetaData;
+import org.eclipse.jetty.http3.qpack.Instruction;
 
-public class TestDecoderHandler implements QpackDecoder.Handler, Instruction.Handler
+public class TestEncoderHandler implements Instruction.Handler
 {
-    private final LinkedList<MetaData> _metadataList = new LinkedList<>();
     private final LinkedList<Instruction> _instructionList = new LinkedList<>();
-
-    @Override
-    public void onMetaData(long streamId, MetaData metadata, boolean wasBlocked)
-    {
-        _metadataList.add(metadata);
-    }
 
     @Override
     public void onInstructions(List<Instruction> instructions)
@@ -43,11 +36,6 @@ public class TestDecoderHandler implements QpackDecoder.Handler, Instruction.Han
         return byteBuffer;
     }
 
-    public MetaData getMetaData()
-    {
-        return _metadataList.poll();
-    }
-
     public Instruction getInstruction()
     {
         return _instructionList.poll();
@@ -55,6 +43,6 @@ public class TestDecoderHandler implements QpackDecoder.Handler, Instruction.Han
 
     public boolean isEmpty()
     {
-        return _metadataList.isEmpty() && _instructionList.isEmpty();
+        return _instructionList.isEmpty();
     }
 }
